+++ R/Rdimtools-package.R
@@ -14,7 +14,8 @@
 #' @import RcppDE
 #' @import maotai
 #' @import ManifoldOptim
-#' @importFrom nabor knn
+#' @importFrom RANN nn2
+#' @importFrom MASS sammon
 #' @importFrom mclustcomp mclustcomp
 #' @importFrom utils packageVersion combn getFromNamespace
 #' @importFrom RSpectra eigs svds