+++ R/zzz.R
@@ -1,5 +1,3 @@
-.pkgenv <- new.env(parent = emptyenv())
-
 .onAttach <- function(...){
   ## Retrieve Year Information
   date <- date()