+++ NEWS.md
@@ -1,6 +1,6 @@
 # Rdimtools 1.0.4
 
-* Two methods for supervised feature selection `do.wdfs` and `do.uwdfs` are added.
+* Methods for supervised feature selection `do.wdfs`, `do.uwdfs`, and `do.mifs` are added.
 * `README` now shows numbers of currently available functions for DR and IDE.
 
 # Rdimtools 1.0.3