+++ NEWS.md
@@ -17,7 +17,7 @@
 
 # Rdimtools 1.0.1
 
-* NEWS reformatted and [package website](https://kyoustat.com/Rdimtools) is now available.
+* NEWS reformatted and [package website](http://kyoustat.com/Rdimtools/) is now available.
 * Dependency on R package [ADMM](https://CRAN.R-project.org/package=ADMM) is removed.